relocate get group test
