+++ sub-packages/bionemo-testing/tests/bionemo/testing/test_megatron_parallel_state_utils.py
@@ -65,6 +65,21 @@ def test_rank7_last_pipeline():
         assert parallel_state.is_pipeline_last_stage()
 
 
+def test_get_pp_group():
+    with megatron_parallel_state_utils.mock_distributed_parallel_state(world_size=2, pipeline_model_parallel_size=2):
+        assert parallel_state.get_pipeline_model_parallel_group() is not None
+
+
+def test_get_tp_group():
+    with megatron_parallel_state_utils.mock_distributed_parallel_state(world_size=2, tensor_model_parallel_size=2):
+        assert parallel_state.get_tensor_model_parallel_group() is not None
+
+
+def test_get_cp_group():
+    with megatron_parallel_state_utils.mock_distributed_parallel_state(world_size=2, context_parallel_size=2):
+        assert parallel_state.get_context_parallel_group() is not None
+
+
 def test_all_reduce():
     # Adapted from https://github.com/pytorch/pytorch/blob/main/test/distributed/test_fake_pg.py
     with megatron_parallel_state_utils.mock_distributed_parallel_state(world_size=2, rank=1):
